unbreak winforms control
When mpv support was added the WinFormMPlayerControl was modified
to no longer automatically initialized a player and needs it passed in.

The documentation was not updated.  Add a new SetPlayer method that can
be used to change the mplayer instance after the control object has
already been initialized.

WinForms Error when playing, following example #22